requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

msgraph-sdk
msgraph-core
azure-identity
azure-functions
azure-functions-durable
azure-storage-blob
azure-search-documents
python-dotenv
llama-index
docx2txt
openpyxl
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
llama-index-vector-stores-azureaisearch
python-dotenv
torch
transformers
python-pptx
Pillow